registry_server/requirements.txt
django
djangorestframework
django-cors-headers
gunicorn
requests
djoser
fastapi
uvicorn
psutil
